refactor: split message list state, rendering, and navigation
